Create a function to submit AddGame form
