+++ src/game/components/AddGame/__tests__/AddGame.test.tsx
@@ -1,6 +1,7 @@
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import AddGame from "./AddGame";
+import AddGame from "../AddGame";
+import { submitAddGame } from "./submitAddGame";
 
 describe("Given the AddGame component", () => {
   const user = userEvent.setup();
@@ -68,26 +69,7 @@ describe("Given the AddGame component", () => {
     test("Then it should show an enabled 'Create game' button", async () => {
       render(<AddGame sendData={mockSendData} />);
 
-      const nameField = screen.getByLabelText(/name/i);
-      const priceField = screen.getByLabelText(/price/i);
-      const rateField = screen.getByLabelText(/rate/i);
-      const descriptionField = screen.getByLabelText(/description/i);
-      const develeportField = screen.getByLabelText(/developer/i);
-      const dateField = screen.getByLabelText(/date/i);
-      const imageUrlField = screen.getByLabelText(/image url/i);
-      const imageAltField = screen.getByLabelText(/alternative text/i);
-
-      await user.type(nameField, "Counter Strike");
-      await user.type(priceField, "3");
-      await user.type(rateField, "2");
-      await user.type(
-        descriptionField,
-        "Un juego en equipo 5 contra 5 de tipo shooter",
-      );
-      await user.type(develeportField, "Valve");
-      await user.type(dateField, "2024-12-29");
-      await user.type(imageUrlField, "counterstrike.webp");
-      await user.type(imageAltField, "counter strike cover");
+      await submitAddGame();
 
       const createGameButton = screen.getByRole("button", {
         name: /create game/i,
